Add Finanzas Personales skill pipeline with per-module selection, and hosting support
- New skill_loader builds a dynamic system prompt from only the requested
  modules (plus always-on base rules), keeping LLM cost proportional to
  what the user actually asked for.
- New financial_snapshots / upload_module_requests tables track per-module
  results and resolved dependencies across periods.
- Worker now parses the full skill JSON contract and upserts snapshots,
  on top of the existing transactions pipeline (unchanged).
- Added /auth/confirm and /auth/resend-code (Cognito sign-up confirmation
  was previously unreachable from the API).
- Fixed Decimal fields serializing as JSON strings in dashboard/transaction/
  account/installment/exchange-rate/category-limit read schemas.
- Fixed CORS headers missing on bare-Exception 500s (ServerErrorMiddleware
  sits outside CORSMiddleware in Starlette's stack).
- Added an in-process worker mode (RUN_WORKER_INLINE) plus a Dockerfile/
  render.yaml for free-tier hosting without a second paid worker service.

+++ app/schemas/account.py
@@ -27,5 +27,5 @@ class AccountRead(BaseModel):
     account_type: AccountType
     institution: str | None
     currency: CurrencyType
-    current_balance: Decimal
+    current_balance: float
     is_active: bool
